# perf: use tantivy to optimize histogram & count without filter