impl #6906

This PR can optimize the query for `histogram` and `count` with none filter.


### Version: `main`

Index size: `4.5GB`

```
select histogram(_timestamp, '1 minute') AS zo_sql_key, count(*) AS zo_sql_num from default GROUP BY zo_sql_key ORDER BY zo_sql_key
```
took: `420ms`

```
select count(*) AS zo_sql_num from default
```
took: `155ms`

### Version: `This PR`

Index size: `4.7GB`

```
select histogram(_timestamp, '1 minute') AS zo_sql_key, count(*) AS zo_sql_num from default GROUP BY zo_sql_key ORDER BY zo_sql_key
```
took: `316ms`

```
select count(*) AS zo_sql_num from default
```
took: `30ms`